requirements.txt
chromadb>=0.4.0
langchain>=0.1.0
pypdf>=3.0.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
unstructured>=0.10.0
markdown>=3.5.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.17.0
altair>=5.0.0